+++ app/crud/catalogs/pathological_background_crud.py
@@ -1,5 +1,5 @@
 from sqlalchemy import case
-from sqlalchemy.exc import SQLAlchemyError
+from sqlalchemy.exc import SQLAlchemyError, IntegrityError
 from sqlalchemy.orm import Session
 from utils.db import db_mapping_rows_to_dict
 
@@ -71,6 +71,16 @@ def update_pathological_background_by_id(
 
 def delete_pathological_back(db: Session, pathological_back_id:int):
     pathological_back = db.query(PathologicalBackground).filter(PathologicalBackground.id==pathological_back_id).first()
-    db.delete(pathological_back)
-    db.commit()
-    return {"status" : True}
+    
+    if pathological_back == None:
+        return None
+    try:
+        db.delete(pathological_back)
+        db.commit()
+    except IntegrityError:
+        db.rollback()
+        return {"status": 2, "msg": "Can not delete pathological background, referenced by other table"}
+    except:
+        db.rollback()
+        return {"status": 3, "msg": "Internal Server Error"}
+    return {"status" : 1, "msg": "Pathological background deleted successfully"}
